Add PostgreSQL/MongoDB and Upgrade Apache

+++ uninstall.sh
@@ -32,7 +32,7 @@ Usage(){
 Usage: $0 [  ${CMSG}all${CEND} | ${CMSG}web${CEND} | ${CMSG}db${CEND} | ${CMSG}php${CEND} | ${CMSG}hhvm${CEND} | ${CMSG}pureftpd${CEND} | ${CMSG}redis${CEND} | ${CMSG}memcached${CEND} ]
 ${CMSG}all${CEND}            --->Uninstall All
 ${CMSG}web${CEND}            --->Uninstall Nginx/Tengine/Apache/Tomcat
-${CMSG}db${CEND}             --->Uninstall MySQL/MariaDB/Percona/AliSQL
+${CMSG}db${CEND}             --->Uninstall MySQL/MariaDB/Percona/AliSQL/PostgreSQL/MongoDB
 ${CMSG}php${CEND}            --->Uninstall PHP
 ${CMSG}hhvm${CEND}           --->Uninstall HHVM
 ${CMSG}pureftpd${CEND}       --->Uninstall PureFtpd
@@ -95,15 +95,42 @@ Print_DB() {
   [ -e "${db_install_dir}" ] && echo "${db_install_dir}"
   [ -e "/etc/init.d/mysqld" ] && echo "/etc/init.d/mysqld"
   [ -e "/etc/my.cnf" ] && echo "/etc/my.cnf"
+  [ -e "${pgsql_install_dir}" ] && echo "${pgsql_install_dir}"
+  [ -e "/etc/init.d/postgresql" ] && echo "/etc/init.d/postgresql"
+  [ -e "${mongo_install_dir}" ] && echo "${mongo_install_dir}"
+  [ -e "/etc/init.d/mongod" ] && echo "/etc/init.d/mongod"
 }
 
 Uninstall_DB() {
-  [ -e "${db_install_dir}" ] && { service mysqld stop > /dev/null 2>&1; rm -rf ${db_install_dir} /etc/init.d/mysqld /etc/my.cnf /etc/ld.so.conf.d/{mysql,mariadb,percona,alisql}*.conf; }
-  id -u mysql >/dev/null 2>&1 ; [ $? -eq 0 ] && userdel mysql
-  [ -e "${db_data_dir}" ] && /bin/mv ${db_data_dir}{,$(date +%Y%m%d%H)}
-  sed -i 's@^dbrootpwd=.*@dbrootpwd=@' ./options.conf
-  sed -i "s@${db_install_dir}/bin:@@" /etc/profile
-  echo "${CMSG}DB uninstall completed${CEND}"
+  # uninstall mysql,mariadb,percona,alisql 
+  if [ -d "${db_install_dir}/support-files" ];then
+    service mysqld stop > /dev/null 2>&1
+    rm -rf ${db_install_dir} /etc/init.d/mysqld /etc/my.cnf /etc/ld.so.conf.d/{mysql,mariadb,percona,alisql}*.conf
+    id -u mysql >/dev/null 2>&1 ; [ $? -eq 0 ] && userdel mysql
+    [ -e "${db_data_dir}" ] && /bin/mv ${db_data_dir}{,$(date +%Y%m%d%H)}
+    sed -i 's@^dbrootpwd=.*@dbrootpwd=@' ./options.conf
+    sed -i "s@${db_install_dir}/bin:@@" /etc/profile
+  fi
+  # uninstall postgresql
+  if [ -e "${pgsql_install_dir}/bin/psql" ]; then
+    service postgresql stop > /dev/null 2>&1
+    rm -rf ${pgsql_install_dir} /etc/init.d/postgresql ${php_install_dir}/etc/php.d/ext-pgsql.ini
+    id -u postgres >/dev/null 2>&1 ; [ $? -eq 0 ] && userdel postgres 
+    [ -e "${pgsql_data_dir}" ] && /bin/mv ${pgsql_data_dir}{,$(date +%Y%m%d%H)}
+    sed -i 's@^dbpostgrespwd=.*@dbpostgrespwd=@' ./options.conf
+    sed -i "s@${pgsql_install_dir}/bin:@@" /etc/profile
+    echo "${CMSG}PostgreSQL uninstall completed${CEND}"
+  fi
+  # uninstall mongodb 
+  if [ -e "${mongo_install_dir}/bin/mongo" ]; then
+    service mongod stop > /dev/null 2>&1
+    rm -rf ${mongo_install_dir} /etc/mongod.conf /etc/init.d/mongod /var/log/mongodb /tmp/mongo*.sock
+    id -u mongod > /dev/null 2>&1 ; [ $? -eq 0 ] && userdel mongod 
+    [ -e "${mongo_data_dir}" ] && /bin/mv ${mongo_data_dir}{,$(date +%Y%m%d%H)}
+    sed -i 's@^dbmongopwd=.*@dbmongopwd=@' ./options.conf
+    sed -i "s@${mongo_install_dir}/bin:@@" /etc/profile
+    echo "${CMSG}MongoDB uninstall completed${CEND}"
+  fi
 }
 
 Print_PHP() {
@@ -177,7 +204,7 @@ Print_curlopenssl() {
 
 Uninstall_curlopenssl() {
   [ -e "/usr/local/bin/curl" ] && rm -rf /usr/local/lib/libcurl* /usr/local/bin/curl
-  [ -d "${openssl_install_dir}" ] && { rm -rf ${openssl_install_dir} /etc/ld.so.conf.d/openssl.conf; ldconfig; }
+  [ -d "${openssl_install_dir}" ] && { rm -rf ${openssl_install_dir} /etc/ld.so.conf.d/z.openssl.conf; ldconfig; }
 }
 
 Menu(){
@@ -186,7 +213,7 @@ while :; do
 What Are You Doing?
 \t${CMSG}0${CEND}. Uninstall All
 \t${CMSG}1${CEND}. Uninstall Nginx/Tengine/Apache/Tomcat
-\t${CMSG}2${CEND}. Uninstall MySQL/MariaDB/Percona/AliSQL
+\t${CMSG}2${CEND}. Uninstall MySQL/MariaDB/Percona/AliSQL/PostgreSQL/MongoDB
 \t${CMSG}3${CEND}. Uninstall PHP
 \t${CMSG}4${CEND}. Uninstall HHVM
 \t${CMSG}5${CEND}. Uninstall PureFtpd
@@ -197,7 +224,7 @@ What Are You Doing?
   echo
   read -p "Please input the correct option: " Number
   if [[ ! $Number =~ ^[0-7,q]$ ]]; then
-    echo "${CWARNING}input error! Please only input 0,1,2,3,4,5,6,7 and q${CEND}"
+    echo "${CWARNING}input error! Please only input 0~7 and q${CEND}"
   else
     case "$Number" in
     0)
